services/iot_service/requirements.txt
paho-mqtt==1.6.1
pytz==2023.3
psycopg2==2.9.6